merge from main eutester into my fork, testing branch

+++ eucaops/__init__.py
@@ -100,12 +100,12 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops, ELBops):
             ### Private cloud with root access 
             ### Need to get credentials for the user if there arent any passed in
             ### Need to create service manager for user if we have an ssh connection and password
+            clc_array = self.get_component_machines("clc")
+            self.clc = clc_array[0]
+            walrus_array = self.get_component_machines("ws")
+            self.walrus = walrus_array[0]
+            self.sftp = self.clc.ssh.connection.open_sftp()
             if self.download_creds:
-                clc_array = self.get_component_machines("clc")
-                self.clc = clc_array[0]
-                walrus_array = self.get_component_machines("ws")
-                self.walrus = walrus_array[0]
-
                 if self.credpath is None:
                     ### TRY TO GET CREDS ON FIRST CLC if it fails try on second listed clc, if that fails weve hit a terminal condition
                     try:
@@ -349,12 +349,14 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops, ELBops):
         detaching = []
         not_exist = []
         line = '\n----------------------------------------------------------------------------------------------------\n'
-        vol_str = volumes or "test_resources['volumes']"
-        self.debug('clean_up_test_volumes starting, volumes:'+str(vol_str))
 
-        volumes = volumes or  self.test_resources['volumes']
+        volumes = volumes or self.test_resources['volumes']
         if not volumes:
+            self.debug('clean_up_test_volumes, no volumes passed to delete')
             return
+        self.debug('clean_up_test_volumes starting\nVolumes to be deleted:' + ",".join(str(x) for x in volumes))
+
+
 
         for vol in volumes:
             try:
@@ -409,7 +411,8 @@ class Eucaops(EC2ops,S3ops,IAMops,STSops,CWops, ASops, ELBops):
             self.monitor_euvolumes_to_status(detaching, status='available', attached_status=None,timeout=timeout)
         self.debug('clean_up_volumes: Deleteing volumes now...')
         self.print_euvolume_list(euvolumes)
-        self.delete_volumes(euvolumes, timeout=timeout)
+        if euvolumes:
+            self.delete_volumes(euvolumes, timeout=timeout)
 
                     
     def get_current_resources(self,verbose=False):
